UIMA-798 Apply Bhavani's patch
git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimacpp/trunk@636443 13f79535-47bb-0310-9956-ffa450edef68

+++ src/framework/annotator_context.cpp
@@ -613,7 +613,7 @@ namespace uima {
   }
 
   void AnnotatorContext::release(vector<std::string*> & returnValues) const {
-    for (int i=0; i < returnValues.size();i++) {
+    for (size_t i=0; i < returnValues.size();i++) {
       delete returnValues.at(i);
     }
     returnValues.clear();
